chore: split off mod litnum and number
